Extract graph output helpers partial
